fix: 840 edit contentModal fixs 🐛
