+++ apps/web/locals/da.json
@@ -1580,6 +1580,11 @@
       "common": {
         "enterAmount": "Indtast beløb",
         "feeNote": "Minimum DKK 8, da Kiibee opkræver DKK 5 i transaktionsgebyr"
+      },
+      "physicalLink": {
+        "title": "Fysisk produkt",
+        "description": "Hvis du har en fysisk version af produktet, kan du linke til en webshop.",
+        "placeholder": "Tilføj produktlink"
       }
     }
   },
